refactor: jdbc mapping으로 dialect 변경에 따른 추상화제거
