--- src/main/java/persistence/sql/ddl/Dialect.java
@@ -1,6 +1,0 @@
-package persistence.sql.ddl;
-
-public interface Dialect {
-
-    String getDialectType(Class<?> type);
-}
