+++ src/test/java/persistence/sql/ddl/JdbcTypeJavaClassMappingTest.java
@@ -1,0 +1,15 @@
+package persistence.sql.ddl;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class JdbcTypeJavaClassMappingTest {
+
+
+    @Test
+    void test() {
+
+    }
+
+}
